# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: WOVSDH Nails Booking API
  description: |
    Complete RESTful API for online nail salon booking system with Telegram notifications, Google Calendar integration, and email confirmations.

    ## Base URLs

    - **Local Development**: `http://localhost:3000`
    - **Production**: `https://your-domain.com`

    ## Authentication

    ### Public Endpoints
    Most endpoints are public and don't require authentication. However, Supabase RLS (Row Level Security) ensures data isolation.

    ### Admin Endpoints
    Admin endpoints require `X-Admin-Secret` header with the value of `ADMIN_SECRET_KEY` environment variable.

    ### Internal Endpoints
    Internal endpoints (confirmed bookings, email sending) require `X-Internal-Secret` header with the value of `ADMIN_SECRET_KEY`.

    ## Response Format

    All responses are JSON. Errors follow a standard format:

    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message"
      }
    }
    ```

    ## Dates and Times

    All timestamps are returned in ISO 8601 format with timezone offset (e.g., `2025-06-01T08:00:00+03:00`).

    All dates in query parameters use format `YYYY-MM-DD`.

    ## Rate Limiting

    No rate limiting is currently implemented. For production use, implement rate limiting via Vercel's edge functions or external service.

    ## Webhooks

...
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/admin/slots:
    get:
      summary: List all slots for admin (with booking info)
      description: |
        Retrieve all slots for a studio within a date range, including information about bookings. Each slot can show the associated booking (if booked).

        Requires admin authentication.
      operationId: getAdminSlots
      tags:
        - Admin
      security:
        - AdminSecret: []
      parameters:
        - name: studio_id
          in: query
          required: true
          schema:
            type: string
            enum:
              - rishon
              - ashdod
          description: Studio identifier
          example: rishon
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Start date (YYYY-MM-DD)
          example: '2025-06-01'
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
          description: End date (YYYY-MM-DD)
          example: '2025-06-30'
      responses:
        '200':
          description: Slots retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAdminSlotsResponse'
          examples:
            success:
              value:
                slots:
                  - id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                    studio_id: rishon
                    start_at: '2025-06-01T08:00:00+03:00'
                    end_at: '2025-06-01T09:00:00+03:00'
                    status: available
                  - id: 7f8c72c0-1234-5678-9abc-def012345678
                    studio_id: rishon
                    start_at: '2025-06-01T09:00:00+03:00'
                    end_at: '2025-06-01T10:00:00+03:00'
                    status: booked
                    booking:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      client_first_name: Анна
                      client_last_name: Иванова
                      status: PENDING
        '401':
          description: Unauthorized - invalid or missing admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                unauthorized:
                  value:
                    error:
                      code: UNAUTHORIZED
                      message: Invalid admin credentials
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      summary: Generate slots for a date range
      description: |
        Create multiple slots for a specific date, studio, and time period with the specified duration.

        System will:
        1. Split the time period into slots of the specified duration
        2. Create slots that don't overlap with existing ones
        3. Return count of created and skipped slots

        Requires admin authentication.

        Example: Create 1-hour slots from 09:00 to 17:30 will generate:
        09:00-10:00, 10:00-11:00, ..., 16:30-17:30
      operationId: generateSlots
      tags:
        - Admin
      security:
        - AdminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSlotsRequest'
            examples:
              example1:
                value:
                  studio_id: rishon
                  date: '2025-06-01'
                  slot_duration_minutes: 60
                  start_time: '09:00'
                  end_time: '17:30'
      responses:
        '200':
          description: Slots generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateSlotsResponse'
          examples:
            success:
              value:
                created: 8
                skipped: 0
                slots:
                  - id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                    studio_id: rishon
                    start_at: '2025-06-01T09:00:00+03:00'
                    end_at: '2025-06-01T10:00:00+03:00'
                    status: available
                  - id: 7f8c72c0-1234-5678-9abc-def012345678
                    studio_id: rishon
                    start_at: '2025-06-01T10:00:00+03:00'
                    end_at: '2025-06-01T11:00:00+03:00'
                    status: available
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/admin/slots/{id}:
    delete:
      summary: Delete a single slot
      description: |
        Delete a specific slot by ID. If the slot is booked, the associated booking will be cancelled and the client notified.

        Requires admin authentication.
      operationId: deleteSlot
      tags:
        - Admin
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Slot ID
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
      responses:
        '200':
          description: Slot deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSlotResponse'
          examples:
            success:
              value:
                message: Slot deleted successfully
                id: f47ac10b-58cc-4372-a567-0e02b2c3d479
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    SlotDTO:
      type: object
      description: A time slot (available, booked, or blocked)
      properties:
        id:
          type: string
          format: uuid
          description: Unique slot identifier
        studio_id:
          type: string
          description: Studio identifier (rishon or ashdod)
        start_at:
          type: string
          format: date-time
          description: Slot start time (ISO 8601 with timezone)
        end_at:
          type: string
          format: date-time
          description: Slot end time (ISO 8601 with timezone)
        status:
          type: string
          enum:
            - available
            - booked
            - blocked
          description: Slot status
      required:
        - id
        - studio_id
        - start_at
        - end_at
        - status
    AdminSlotDTO:
      allOf:
        - $ref: '#/components/schemas/SlotDTO'
        - type: object
          properties:
            booking:
              type: object
              description: Booking information (only if slot is booked)
              properties:
                id:
                  type: string
                  format: uuid
                client_first_name:
                  type: string
                client_last_name:
                  type: string
                status:
                  type: string
                  enum:
                    - PENDING
                    - CONFIRMED
                    - CANCELLED
...
    GenerateSlotsRequest:
      type: object
      description: Request to generate slots for a date
      properties:
        studio_id:
          type: string
          description: Studio identifier
        date:
          type: string
          format: date
          description: Date in YYYY-MM-DD format
        slot_duration_minutes:
          type: integer
          description: Duration of each slot in minutes (e.g., 60)
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Day start time in HH:mm format (local time)
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Day end time in HH:mm format (local time)
      required:
        - studio_id
        - date
        - slot_duration_minutes
        - start_time
        - end_time
...
    GetAdminSlotsResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/AdminSlotDTO'
    GenerateSlotsResponse:
      type: object
      properties:
        created:
          type: integer
          description: Number of slots created
        skipped:
          type: integer
          description: Number of slots skipped (already existed)
        slots:
          type: array
          items:
            $ref: '#/components/schemas/SlotDTO'
    DeleteSlotResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: string
          format: uuid
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_PARAMS
                - SLOT_NOT_FOUND
                - SLOT_UNAVAILABLE
                - SLOT_HAS_ACTIVE_BOOKING
                - BOOKING_NOT_FOUND
                - ALREADY_CANCELLED
                - UNAUTHORIZED
                - VALIDATION_ERROR
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
...
  securitySchemes:
    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret key (value of ADMIN_SECRET_KEY env var)

    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Internal secret key (value of ADMIN_SECRET_KEY env var)
...